requirements.txt
flask>=2.0.0
redis>=4.0.0
psycopg2-binary>=2.8.0
requests>=2.25.0